Filter empty responses
The server will only read from stdout when it requires a value, not
after each command. Because of this, writing `""` to stdout for commands
that require no response will cause the server to be out of synch with
the client.

This filters out `""` from the response so that only commands are
written to stdout.